+++ apps/content-blocks-sandbox/config/bundles.php
@@ -8,6 +8,7 @@ return [
     Symfony\UX\TwigComponent\TwigComponentBundle::class => ['all' => true],
     Symfony\UX\LiveComponent\LiveComponentBundle::class => ['all' => true],
     Symfony\UX\StimulusBundle\StimulusBundle::class => ['all' => true],
+    Liip\ImagineBundle\LiipImagineBundle::class => ['all' => true],
     ContentBlocks\ContentBlocksBundle::class => ['all' => true],
     ContentBlocks\Kit\ContentBlocksKitBundle::class => ['all' => true],
 ];